chore: refactor DoH resolver
